Initial SDL3 and WGPU integration (hardcoded triangle).

+++ Ion/Ion.Extensions.Assets.Abstractions/IAsset.cs
@@ -3,6 +3,6 @@ namespace Ion.Extensions.Assets;
 
 public interface IAsset : IDisposable
 {
-	int Id { get; }
+	nint Id { get; }
 	string Name { get; }
 }
