+++ Ion/Ion.Extensions.Assets.Abstractions/IAssetManager.cs
@@ -5,7 +5,7 @@ public interface IBaseAssetManager
 {
 	IAssetLoader GetLoader(Type assetType);
 	T Set<T>(T asset) where T : class, IAsset;
-	T? Get<T>(int id) where T : class, IAsset;
+	T? Get<T>(nint id) where T : class, IAsset;
 	void Unload<T>(T asset) where T : class, IAsset;
 }
